refactor(export): move grid translation helpers to translators
Value conversion belongs in translators next to translateGridTracks
and translateGridCells; the transformer partial keeps only the
fallback policy and degradation reporting.
